+++ app/policies/user_policy.rb
@@ -1,0 +1,52 @@
+class UserPolicy < Policy
+  module Collection
+    def index?
+      true
+    end
+
+    def deactivated?
+      index?
+    end
+
+    def new?
+      user.has_role?(:users) || !User.any?
+    end
+
+    def create?
+      new?
+    end
+    def create_first?
+      new?
+    end
+
+    def manage?
+      user.has_role?(:users)
+    end
+  end
+
+  module Member
+    def edit?
+      user == record || user.has_role?(:users)
+    end
+
+    def show?
+      edit?
+    end
+
+    def delete_image?
+      edit?
+    end
+
+    def update_openid?
+      edit?
+    end
+
+    def policies?
+      user.has_role?(:users)
+    end
+
+    def destroy?
+      user.has_role?(:users)
+    end
+  end
+end